Separate styles in VideoGallery.styles.ts
