+++ packages/react-components/src/components/VideoGallery/styles/FloatingLocalVideoLayout.styles.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { IStackStyles } from '@fluentui/react';
+import { IStackStyles, IStyle } from '@fluentui/react';
 
 /**
  * @private
@@ -15,4 +15,18 @@ export const rootLayoutStyle: IStackStyles = {
  */
 export const innerLayoutStyle: IStackStyles = {
   root: { position: 'relative', height: '100%', width: '100%', padding: '0.5rem' }
+};
+
+/**
+ * @private
+ */
+export const layerHostStyle: IStyle = {
+  position: 'absolute',
+  left: 0,
+  top: 0,
+  width: '100%',
+  height: '100%',
+  overflow: 'hidden',
+  // pointer events for layerHost set to none to make descendants interactive
+  pointerEvents: 'none'
 };
